SkillBase.cs:  removed SpellLine.Equals - method was poorly defined, never referenced, and missing the accompanying override to GetHashCode Util.cs:  added try/finally for suspended thread to ensure it gets resumed DetailDisplayHandler.cs:  only change is line 341 - cast bitwise-or operand to ushort PacketProcessor.cs:  added try/finally for suspended thread to ensure it gets resumed
git-svn-id: https://svn.code.sf.net/p/dolserver/code/DOLSharp/trunk@1522 c9877f0d-d2ea-46ae-93cb-121acdf3e971

+++ GameServer/packets/Client/168/DetailDisplayHandler.cs
@@ -338,7 +338,7 @@ namespace DOL.GS.PacketHandler.Client.v168
 						}
 						else if (objectType == 19)
 						{
-							ushort questID = (ushort)((unk_186 << 12) | (objectID >> 4));
+							ushort questID = (ushort)((unk_186 << 12) | (ushort)(objectID >> 4));
 							int index = objectID & 0x0F;
 							if (questID == 0)
 								return 1; // questID == 0, wrong ID ?
